requirements.txt
# Core
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=8.0.0
pytest-bdd>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# Security
bandit>=1.7.0
safety>=3.0.0
pip-audit>=2.7.0

# Linting
ruff>=0.1.0

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# Telegram Bot
aiogram>=3.0.0

# Database
sqlalchemy>=2.0.0
ydb>=3.0.0

# Object Storage (S3)
boto3>=1.34.0
